requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements.in
#
anyio==3.6.2
    # via httpcore
attrs==22.1.0
    # via jsonschema
boto3==1.26.27
    # via -r ./requirements.in
botocore==1.29.27
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==3.0.4
    # via requests
charmed-kubeflow-chisme==0.0.5
    # via -r ./requirements.in
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
h11==0.14.0
    # via httpcore
httpcore==0.16.2
    # via httpx
httpx==0.23.1
    # via lightkube
idna==2.10
    # via
    #   anyio
    #   requests
    #   rfc3986
jinja2==3.1.2
    # via charmed-kubeflow-chisme
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via serialized-data-interface
lightkube==0.11.0
    # via
    #   -r ./requirements.in
    #   charmed-kubeflow-chisme
lightkube-models==1.25.4.4
    # via
    #   -r ./requirements.in
    #   lightkube
markupsafe==2.1.1
    # via jinja2
oci-image==1.0.0
    # via -r ./requirements.in
ops==1.5.4
    # via
    #   -r ./requirements.in
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
pyrsistent==0.19.2
    # via jsonschema
python-dateutil==2.8.2
    # via botocore
pyyaml==5.4
    # via
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.25.0
    # via serialized-data-interface
rfc3986[idna2008]==1.5.0
    # via httpx
ruamel-yaml==0.17.21
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
s3transfer==0.6.0
    # via boto3
serialized-data-interface==0.4.0
    # via -r ./requirements.in
six==1.16.0
    # via
    #   jsonschema
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tenacity==8.1.0
    # via -r ./requirements.in
urllib3==1.26.13
    # via
    #   botocore
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools